Feature/transactions

+++ src/main/java/com/md/demo/service/ItemServiceImpl.java
@@ -1,5 +1,6 @@
 package com.md.demo.service;
 
+import com.md.demo.exception.NoSuchItemExistException;
 import com.md.demo.model.Item;
 import com.md.demo.repository.ItemRepository;
 import org.slf4j.Logger;
@@ -23,15 +24,26 @@ public class ItemServiceImpl implements ItemService {
 		this.itemRepository = requireNonNull(itemRepository, "itemRepository can not be null");
 	}
 
-	@Transactional(isolation = Isolation.READ_COMMITTED)
+	@Override
+	@Transactional(isolation = Isolation.READ_COMMITTED, readOnly = true)
 	public Item getItemById(Integer id) {
 		LOGGER.info("About to find Item from DB by id: " + id);
 		Optional<Item> byId = itemRepository.findById(id);
-		return byId.orElse(null);
+
+		return byId.orElseThrow(() -> new NoSuchItemExistException("No such item in DB"));
 	}
 
+	@Override
 	@Transactional(isolation = Isolation.READ_COMMITTED)
-	public void deleteItem(Integer id) {
-		itemRepository.deleteById(id);
+	public boolean isItemDeleted(Integer id) {
+		boolean itemIsDeleted = false;
+		Item itemById = getItemById(id);
+		LOGGER.info("About to delete an Item from DB by id: " + id);
+		if (itemById.getAverageRating() != null) {
+			itemRepository.deleteById(id);
+			return itemIsDeleted = true;
+		} else {
+			throw  new NoSuchItemExistException("No such Item in DB!");
+		}
 	}
 }
